backend/requirements.txt
asgiref==3.6.0
Django==3.2.19
django-environ==0.10.0
django-extra-fields==3.0.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.9.0
djoser==2.1.0
gunicorn==20.0.4
oauthlib==3.2.2
paramiko==2.12.0
passlib==1.7.2
pgadmin4==4.29
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2020.5
requests==2.30.0
requests-oauthlib==1.3.1
simplejson==3.16.0
six==1.16.0